+++ src/main.ts
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
-import { app } from "./app";
+
+import { app } from "app.js";
 
 app();