+++ src/modules/depedenciesInstaller/useCases/installDependencies/index.ts
@@ -1,3 +1,0 @@
-import { InstallDependencies } from "./InstallDependencies";
-
-export default InstallDependencies;