+++ src/modules/questionnaire/errors/QuestionnaireError/index.ts
@@ -1,7 +1,0 @@
-import { UseCaseError } from "@/shared/core/errors/ErrorUseCase";
-
-export class QuestionnaireError extends Error implements UseCaseError {
-  constructor() {
-    super("An error occurred during the questionnaire process.");
-  }
-}