+++ src/modules/questionnaire/useCases/Questionnaire.ts
@@ -1,5 +1,5 @@
-import { UseCase } from "@/shared/core/modules/UseCase";
-import { QuestionnaireRepository } from "../../repositories/contracts/QuestionnaireRepository";
+import { UseCase } from "@shared/core/modules/UseCase.js";
+import { QuestionnaireRepository } from "../repositories/contracts/QuestionnaireRepository.js";
 
 export class Questionnaire implements UseCase<null> {
   constructor(private questionnaireRepository: QuestionnaireRepository) {}