+++ src/modules/setupManager/useCases/environmentSettings/EnvironmentSettings.ts
@@ -1,7 +1,7 @@
-import { UseCase } from "@/shared/core/modules/UseCase";
-import chalk from "chalk";
-import { createSpinner } from "nanospinner";
-import { SetupManagerRepository } from "../../repositories/contracts/SetupManagerRepository";
+import { SetupManagerRepository } from '@modules/setupManager/repositories/contracts/SetupManagerRepository.js';
+import { UseCase } from '@shared/core/modules/UseCase.js';
+import chalk from 'chalk';
+import { createSpinner } from 'nanospinner';
 
 export class EnvironmentSettings implements UseCase<Record<string, string>> {
   constructor(private setupManagerRepository: SetupManagerRepository) {}
@@ -11,13 +11,13 @@ export class EnvironmentSettings implements UseCase<Record<string, string>> {
       `We're rushing to setup your project!`,
     ).start();
     try {
-      if (answers.stack === "Frontend") {
+      if (answers.stack === 'Frontend') {
         await this.setupManagerRepository.setupFrontendConfigurations(answers);
-        spinner.success({ text: "Project environment is now complete!" });
+        spinner.success({ text: 'Project environment is now complete!' });
         return;
       }
       await this.setupManagerRepository.setupBackendConfigurations(answers);
-      spinner.success({ text: "Project environment is now complete!" });
+      spinner.success({ text: 'Project environment is now complete!' });
     } catch (error) {
       spinner.error({
         text: chalk.red(error),