+++ src/modules/setupManager/useCases/installDependencies/installDependencies.ts
@@ -1,7 +1,7 @@
-import { UseCase } from "@/shared/core/modules/UseCase";
-import chalk from "chalk";
-import { createSpinner } from "nanospinner";
-import { SetupManagerRepository } from "../../repositories/contracts/SetupManagerRepository";
+import chalk from 'chalk';
+import { createSpinner } from 'nanospinner';
+import { UseCase } from 'shared/core/modules/UseCase.js';
+import { SetupManagerRepository } from '../../repositories/contracts/SetupManagerRepository.js';
 
 export class InstallDependencies implements UseCase<Record<string, string>> {
   constructor(private setupManagerRepository: SetupManagerRepository) {}
@@ -12,10 +12,10 @@ export class InstallDependencies implements UseCase<Record<string, string>> {
     ).start();
     try {
       await this.setupManagerRepository.installDependencies(answers);
-      spinner.success({ text: "Project dependencies installed successfully!" });
+      spinner.success({ text: 'Project dependencies installed successfully!' });
       spinner.success({
         text: `${chalk.bold.magenta(
-          "Thank you for using Confectus! See you in the next code!",
+          'Thank you for using Confectus! See you in the next code!',
         )}
       `,
       });