+++ src/modules/templatesManager/errors/FileCopyError/index.ts
@@ -1,7 +1,0 @@
-import { UseCaseError } from "@/shared/core/errors/ErrorUseCase";
-
-export class FileCopyError extends Error implements UseCaseError {
-  constructor() {
-    super("An error occurred during the files creation process. Sorry!");
-  }
-}