+++ src/modules/templatesManager/repositories/implementations/TemplatesManagerRepositoryImplementations.ts
@@ -1,8 +1,8 @@
-import fs from "fs-extra";
-import path from "path";
-import { fileURLToPath } from "url";
-import { FileCopyError } from "../../errors/FileCopyError";
-import { TemplatesManagerRepository } from "../contracts/TemplatesManagerRepository";
+import { FileCopyError } from '@modules/templatesManager/errors/FileCopyError.js';
+import fs from 'fs-extra';
+import path from 'path';
+import { fileURLToPath } from 'url';
+import { TemplatesManagerRepository } from '../contracts/TemplatesManagerRepository.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -14,13 +14,13 @@ export class TemplatesManagerRepositoryImplementations
     templateSource: string[],
     templateDestination: string,
   ): Promise<void> {
-    const isDevelopment = process.env.NODE_ENV === "development";
+    const isDevelopment = process.env.NODE_ENV === 'development';
     const rootPath = isDevelopment
-      ? path.resolve(__dirname, "../../../../")
+      ? path.resolve(__dirname, '../../../../')
       : __dirname;
 
     const templatesPath = (...subpaths: string[]) =>
-      path.join(rootPath, "templates", ...subpaths);
+      path.join(rootPath, 'templates', ...subpaths);
 
     const copyFiles = async (
       source: string,
@@ -45,7 +45,7 @@ export class TemplatesManagerRepositoryImplementations
     await copyFiles(
       templatesPath(...templateSource),
       isDevelopment
-        ? path.resolve("mock", templateDestination)
+        ? path.resolve('mock', templateDestination)
         : templateDestination,
       new FileCopyError(),
     );