+++ src/modules/templatesManager/useCases/handleTemplate/index.ts
@@ -1,3 +1,0 @@
-import { HandleTemplate } from "./handleTemplate";
-
-export default HandleTemplate;