+++ src/types/setting.ts
@@ -1,7 +1,7 @@
-export interface SettingsProps {
+export type SettingsProps = {
   [key: string]: {
     configFiles: { configFileName: string; configFilePath: string[] };
     dependencies: string | null;
     devDependencies: string | null;
   };
-}
+};