+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # confectus
 
+## 1.4.6
+
+### Patch Changes
+
+- Fix eslint configuration for vue.js using Typescript
+
+## 1.4.5
+
+### Patch Changes
+
+- Fix package error that causes erros when was installing depedencies
+
+## 1.4.4
+
+### Patch Changes
+
+- Change configuration of vanilla projects
+
+## 1.4.3
+
+### Patch Changes
+
+- Fix critical error
+
+## 1.4.2
+
+### Patch Changes
+
+- Fix some bugs and errors
+
+## 1.4.1
+
+### Patch Changes
+
+- Fix some bugs and errors
+
 ## 1.4.0
 
 ### Minor Changes